Reduce section.bin cache size with memory optimizations
  - Move EpdFontStyle enum inside EpdFontFamily class as uint8_t
  - Use uint16_t for viewport dimensions, page counts, and element counts
  - Bump section file version to 8 (forces cache regeneration)

+++ lib/Epub/Epub/Page.cpp
@@ -33,7 +33,7 @@ void Page::render(GfxRenderer& renderer, const int fontId, const int xOffset, co
 }
 
 bool Page::serialize(FsFile& file) const {
-  const uint32_t count = elements.size();
+  const uint16_t count = elements.size();
   serialization::writePod(file, count);
 
   for (const auto& el : elements) {
@@ -50,10 +50,10 @@ bool Page::serialize(FsFile& file) const {
 std::unique_ptr<Page> Page::deserialize(FsFile& file) {
   auto page = std::unique_ptr<Page>(new Page());
 
-  uint32_t count;
+  uint16_t count;
   serialization::readPod(file, count);
 
-  for (uint32_t i = 0; i < count; i++) {
+  for (uint16_t i = 0; i < count; i++) {
     uint8_t tag;
     serialization::readPod(file, tag);
